OpenGL/MaskedIcon: use alpha blending (#3124)
Instead of using GL_OR and GL_AND to render transparent icons, keep
the alpha channel in the PNG files, and pass it on to OpenGL.  Then
use GL_BLEND to apply it.

This is faster, notably on Vivante GPUs, which are extremely slow at
GL_COLOR_LOGIC_OP (TRAC #3124).