Event/Kobo/Queue: move RotatePointer instance into class MergeMouse
Apply the rotation before the bounds checks.  There was a big Kobo
regression in portrait mode: the bounds were applied before the mouse
position was rotated, and thus we had everything clipped at y<600
instead of y<800.

As a side effect, this feature is now available on all "console"
targets.  It will allows us to rotate the screen even if the graphics
driver doesn't support it.