Topography/XShape: store floating point offset to map center
Make ShapePoint use "float" (compatible with GL_FLOAT) and store
offsets relative to the map center instead of relative to the shape
center.  This means all XShapes live in the same coordinate system,
and they all use the same transformation matrix.  It will allow using
a single OpenGL draw command for all shapes in a file.