Replacement for "ld -b binary" resource embedding mechanism
"ld -b binary" is not supported by all linkers, particularly not by the
Apple ld64 linker. Now, a small (and portable) assembler program is
generated compiled by the EmbedFileInObjectFile.pm perl module to
generate .o files which are compatible with those compiled with "ld -b
binary". The .mk include order is changed so that the AS variable
(defined in compile.mk) is correctly initialized for resource.mk and
libdata.mk.