Revise jas_stream_tmpfile()
The previous implementation of jas_stream_tmpfile() would always have
failed for WIN32 and Android, so disable the function completely for
these platforms and always return 0. Using tmpnam() causes a compiler
warning when using Apple SDKs, and is also not recommended for Linux. So
use mkstemp() for these targets instead.