Android: Raise target Android SDK Version to 22
This fixes #3491 (cropped display in Immersive Full Screen Mode). Since
API Level 21, there is a new API for implementing a TransitionManager
for a Window. This API is not used in XCSoar, but it seems that it is
used somehow by the OS when the App has target platform 21 or newer, and
the bug does not occur then.
The NDK platform versions remains untouched for now. There is no
"android-22" target in the current NDK. And using NDK target
"android-21" results in code which is not backwards compatible, see
https://code.google.com/p/android-developer-preview/issues/detail?id=168