Dialogs/Task/Manager: remove "Target" button
The "Target" button is misplaced here.  It does not belong here, as it
is a tool to edit the target while flying, and the task manager is for
pre-flight planning.  More confusing is that it does not show the
pending modifications the user just applied in the task editor.  Let's
remove it - this feature should better be called from the "Nav" menu.