Computer/Wind: remove option "use_external_wind"
This option was confusing, because the implementation did not match
the documentation.  If this was enabled, setting the wind manually
would always fail.  Being disabled, this always ignored the external
wind.

With this patch, if the user chooses to let XCSoar calculate the wind,
and a result is available, use that.  Use the external wind only if
auto wind is disabled.

There's no point in letting an external device override XCSoar's
calculated wind.