Screen/Layout: use whole display size for "small screen" formula
Fixes fonts that are too small when the XCSoar main window is small,
because the "small screen" formula determined this is a screen smaller
than 5 inch.  That is bogus.  The formula was designed to detect small
mobile devices that are usually closer to the pilot's eye, which a
desktop computer with a small XCSoar window is not.