Data/Input: remove MC from menu, replace with System/...
The MacCready value should be configured with the InfoBox.  Doing it
in the menu without an InfoBox means you can't see the value being
edited, and this is not a good use case anyway.

The 3 menu item slots get replaced with more important configuration
menu items from Config2: "System", "Plane" and "Devices".