Math/fixed: remove option FIXED_MATH
Now that we have dropped support for WinCE, all relevant devices to
run XCSoar on have a floating-point unit, and there is no need anymore
for the "fixed" class.

(The only remaining target without FPU is Android for ARMv6, but that
is not relevant enough, because nobody has an ARMv6 Android anymore
these days.)

This is the first commit for the transition; it only removes the
FIXED_MATH code.