New Screen Layout Option "Show Menu Button"
The new Option "Show Menu Button" is located in menu
"Config/System/Look/Screen Layout". It is shown only at the KOBO device and
consists of 2 states "on" and "off". When "on" (the default) the menu button
is shown in the upper right corner of the map window. When "off" the
menubutton disapears. The changed setting becomes only active after XCSoar
restart.