Makefile: Remove DARWIN_TOOLCHAIN
The toolchain built with DarwinToolchainBuilder is no longer used. For
cross-compiling for iOS / OS X on non-OSX platforms, the LLVM toolchain
and the ld64 linker need to be in the current binary search path (PATH
environment variable). And the path to a iOS / OS X SDK needs to be
specified using the DARWIN_SDK build variable.