screen layout with 12 infoboxes on the left, vario+3 infoboxes on right
The goal of this patch is to add a screen layout that allows
displaying more info boxes on the screen with compromising the visual
feel of the center map area and with preserving infobox order. This
patch adds a new screen layout in landscape mode where you have 2
columns of 6 infoboxes on the left of the screen and a column with a
vario and 3 more infoboxes on the right. So we get 6 more info boxes
on the left side of the screen in a second column. This new column
also has all the new additional infoboxes (in term of infoxbox
ids). That way the order of info boxes is maintained across layouts
(when you switch layouts) so that first 9 infoboxes stay in the same
place. The 2 left columns are each slightly narrower than the right
column thus providing a balanced look and still a nice center map area
that doesn't feel narrow.