WaypointIconRenderer: add Reachability::Invalid
Commit 3016bf2 added that enum value to
WaypointRenderer::Reachability.  The problem is that there is an
illegal cast from that enum to WaypointIconRenderer::Reachability,
added by commit 3f6cac84, and this bad cast requires us to keep both
enums in sync.  This is error prone and the cast should be removed,
replaced by a common enum.  For now, I'm going to add the same enum
value to work around the regression.