Makefile: download Boost for all targets, including Linux
Our Boost compile-time dependency has caused a lot of trouble for
people trying to compile XCSoar, because many distributions ship
outdated Boost versions which are not compatible with XCSoar.  And te
current stable Boost release 1.60 contains serious bugs in
boost::geometry, which will be fixed by migrating to 1.61 beta.
That's only possible by enforcing the use of this version.