Android/BluetoothGattClientPort: activate Notifications for other modules than HM-10
i have developed a dedicated FLARM device which uses a BLE module
(Laird BL-600).

I made tests in the past with a HM-10 module and now with this BL-600
module. I have made the module completely compatible to the HM-10
(Services, Charcteristics, Attributes...). Everything Looks nearly
good, but i have a Problem on Android Version of XCSoar. It is not
possible to get the notifications be enabled.

I have discovered the Bluetoot LE Api and found, that for activation
of notifications two things are necessary:

* setCharacteristicNotification
* write the value to the descriptor

In your BluetoothGATClientPort.java i think you make only the first
part.

The HM-10 modules works with this, but it looks that other modules
have problems with this missing step.

You can find Information of these two steps in the Android SDK, there
it is documented:

 http://developer.android.com/guide/topics/connectivity/bluetooth-le.html#notification