TraceHistoryRenderer/Trace: use vario scale
The problem with the old system was that the scale changed exclusively based on current data,
so the user has a poor understanding/feel of the range.

Now, the scale is determined from the 10%,90% percentiles of encountered climb rates.
So, on weak days we still get a narrow range, on strong days the gain adjusts appropriately.