Audio: new component "PCMResourcePlayer" to play sounds using PCMPlayer
For targets which have a PCMPlayer implementation, but no native support
for playing sound resources, the PCMResourcePlayer can do this job now.
To achieve this, the sounds are embedded as resources in raw 16 bit
(signed) mono 44,1 kHz PCM format, and the PCMResourcePlayer can
playback these using PCMPlayer.

This adds support for playing event sounds for all targets which use SDL
or ALSA for audio playback.