Audio: Refactoring: introduce new interface "PCMDataSource"
PCMPlayer now reads its PCM data from a PCMDataSource, instead of a
PCMSynthesiser (with the exception of the Android implementation, which
does not support the new interface yet). A PCMDataSource is a more
complex interface, specifies its supported sample rate and enianness,
and delivers data which can have an end. PCMSynthesiser is still the
base class for synthesisers, and is now an abstract subclass of
PCMDataSource.