Audio: add PCM mixer layer
For audio backends where the underlying audio devices can be not be
opened multiple times (or where it is at least not safe to assume they
can), this adds a layer between the audio backend and our PCMPlayer
clients, to mix multiple audio sources (PCMDataSource objects) together.

The class PCMMixer can playback multiple PCM audio streams at the same
time with one PCMPlayer as a backend. The class MixerPCMPlayer provides
a PCMPlayer implementation for this.

This feature is used for the SDL and ALSA audio backends.