Makefile: set TARGET_IS_LINUX=y / TARGET_IS_DARWIN=n correctly when cross compiling for Linux
Until now, the variables TARGET_IS_LINUX was set for Unix targets except
Darwin when compiling on Linux, and TARGET_IS_DARWIN when compiling for
Unix on Darwin. This is not always correct, because it makes sense to
cross-compile for Linux on Darwin (e. g. for the KOBO target). Instead,
insure that TARGET_IS_LINUX is set to "y" for all Linux specific targets
(TARGET_IS_DARWIN was already explicitly set to "y" for all Darwin
targets).