doc/manual/dev: update instructions for installing the Android SDK
The Android SDK is now shipped with the Android Studio IDE, and SDK
components can be installed in this IDE's SDK manager.

The "android" tool can no longer be used in recend Android SDK releases,
but there is now the "sdkmanager" tool for installing SDK components
from the command line.