Waypoint/WaypointReaderSeeYou: handle file with rwwidth field
Some SeeYou WP files contain a field for runway width (called "rwwidth" in the file header, if present). For example, at least some versions of SeeYou Mobile create files like this (if saved due to user editing of a WP). This extra field is inserted just before the radio frequency field, thus shifting that field and the description field to the right. Without this change, when XCSoar uses such a file, the radio frequency isn't shown (in XCSoar's frequency field), the description/comments field in XCSoar contains only the radio frequency (from the WP file), and the description/comments from the WP file are nowhere in XCSoar.