build/python/libs: pass "default-rpath" to glibc build
glibc's Makefile defaults to the configured "--prefix" for the
dlopen() search path, which we however set to
"output/KOBO/lib/arm-linux-gnueabihf/root".  This directory however
doesn't exist on the Kobo, breaking NSS modules, which in turn breaks
DNS access (TRAC #3869).

However there's this dirty hack: pass "default-rpath=/opt/xcsoar/lib"
to the glibc Makefile, which overrides this default, without
correcting the "--prefix".

Another approach would be to use "--prefix=/opt/xcsoar" and pass
DESTDIR to the "install" stage, but that comes with other problems
which I'm not in the mood to solve right now.