build/thirdparty: specify two host triplets instead of one
This is for targets for which a toolchain is used, which was originally
intended for another target ABI, e. g. when the arm-linux-gnueabihf
toolchain used to compile for musl instead of glibc. In this case, we
specfiy "arm-linux-gnueabihf" as toolchain host triplet, and something
like "arm-linux-musleabihf" as actual host triplet.