Moved the download directory out of the output directory.
Every time a 'make clean' or 'rm -rf output' gets run, the
output/download directory gets clobbered. This is pretty annoying,
if one does not have the luxury of a high speed Internet connection.

This patch fixes this problem, by moving the location of the download
directory to the XCSoar root directory.

In addition, the 'clean-download' make target was added, for the case
where one really wants to get rid of the download directory for some
reason.