Increase COMM_MAX_PAYLOAD_SIZE to 2kB
- Reflect change in communication protocol introduced by IMI Gliding

- E-mail correspondence between XCSoar contributor kicune and Juraj Rojko <juraj@rojko.cz> of IMI Gliding in January 2018 confirms change in protocol on IMI side.

Juraj Rojko:
The driver for Erixx was written quite a long time ago, basicaly without our participation, I just provided them with data structures, some basic description and IIRC also with some simple demo code. Since then there were some minor changes in the Erixx FW. COMM_MAX_PAYLOAD_SIZE has now the value of 2048. I am not absolutely sure if that was the only change. If you can recompile the XCSoar, I would recommend to try the new size. If this does not help, then we could compare the definitions in XCSoar with ours. I am sure that the differences are not significant and it should be simple to make it work.