Android: Extend the build system for Andoid Legacy library build
Add parameters ANDROID_LEGACY and ANDROID_INCLUDE_LEGACY_LIB.
Both parameters can be set to 'y' only for targets ANDROID7 and ANDROID86.
64-bit targets are not supported because they did not exist for those lower
API levels.

ANDROID_LEGACY=y builds the native library libxcsoar_legacy.so.
It is build based on API 16 libraries, but compatible and tested down
to API level 12 (Honeycomb, Android 3.1).
The suffix _LEGACY is added to the build directory in the subdirectory output.

ANDROID_INCLUDE_LEGACY_LIB=y runs make again with ANDROID_LEGACY=y to build
libxcsoar_legacy.so. This library is added to the own libxcsoar.so
library in the final APK.
This parameter is automatically set to 'y' when DEBUG=n
This parameter is also supported for the target ANDROIDFAT.
It adds the legacy library to the ANDROID7 and ANDROID86 architectures.

Android/NativeLeScanCallback.cpp is compiled for the standard version only.
Android/posix_memalign.cpp is compiled for the legacy library only to implement a missing function in older Bionic versions.

Android: Make the ANDROIDFAT target work