Thirdpary: Make toolchain a member of Project, and elimiate toolchain as paramter
Move the global variables in libs.py into a class Libs.
Declare libs as Libs in the main module, and pass toolchain through it.
Initialize the lib objects with 'toolchain'
Pass 'toolchain' through Autotools, and MakeProject to Project where it
is stored with the __init__ routine.
Remove the parameter 'toolchain' in all decendents of Project and
replace the use of it with sefl.toolchain.

These changes serve to make all system information and particularly path
information immediately available in all Project entities.
This is required to determine download, shared source and build paths in
the constructor.
Finally the tasks of
- determining the paths
- check routines (refined) is_installed, is_unpacked and is_downloaded
- downloading and unpacking
are clearly separated which are currently mangled into the download and
unpack routines.