Thirdparty: Untangle download, unpack, and build.
To separate download and unpacking of shared sources from building make
downloading tarballs, unpack/untar, and actual building of the libs indenpendent.
Pre-determine all relevant paths and tarball names in the constructor of
class Project. Remove calculation and returning these paths from the unpack
and download routines.
Cache the downloaded and unpacked status in class Project.
Add explicit check_unpacked and check_downloaded functions to Project.
Make the download and unpack routines unconditional. Checking if download and/or
unpacking is required is checked elsewhere.
For each unpack there is now a stamp file to allow for clear dependency tracking
of the downloaded tarball and the unpacked sources.
The dependency of the 'installed' indicator file is now against the unpack stamp file, not the tarball.
Remove the unpack and make_build_path calls from the overridden build() methods
of the lib classes. All preparation of building libs is now handled outside
and before the build call, thus elimiating the nasty dependencies in sub-classes of MakeProject or AutotoolsProject.
However all external interfaces, i.e. thirdparty.py and download.py still work
for the caller as before.
Add a target TEST in libs.py. This target serves as test of building all available
libs. It is only used when thridparty.py is called manually for testing.