CI: Add gitlab-CI
This commit adds support for gitlab-CI in order to autmatically build
XCSoar for several targets and provide the build artifact for testers.

Following targets are supported:
* ANDROID
* UNIX
* PC
* KOBO
* DOCS

Additionally it runs `make check` before building the UNIX target.

In order to run the gitlab-CI, the main xcsoar git repository gets
mirrored by a gitlab repository (this is a service provided by any
gitlab instance). As soon as new commits are available on the master,
the gitlab instance starts a CI "pipeline". This pipeline consists
of several jobs as specified in ".gitlab-ci.yml".

The first job 'prepare-image' builds a docker image containing a build
environment for all supported targets. This docker image is uploaded
to the docker-registry provided by the gitlab server.

Then, for each supported target a separate job runs the xcsoar build
in docker containers based on the previously generated image.

Finally the build artifacts (`./output/`) are uploaded to the
gitlab-server (where they are available for download for 4 weeks).