Doc: Table of contents : titles in sans serif, and spacing
Even when the documents using the xcsoar.sty file where producing text in sans serif font (see PR #267), the titles in the table of content was in sans serif. For example: https://download.xcsoar.org/releases/6.8.11/XCSoar-manual.pdf .
Moreover, the section numbers were sometimes touching the section titles. 

With this PR, the titles in the table of contents uses the same font style (i.e. sans serif) as the rest of the document, and there is always a space between the section number and title.