Should scale the UI to fit the full screen on Android devices with aspect ratio's greater than 16:9 (e.g. Pixel 4 XL)
Currently on Android devices with aspect ratio's greater than 16:9,
XCSoar does not use the full screen and instead will leave a black bar
of unused space at the bottom.

As per this documentation, this should ensure the UI scales to fit the
full screen:

https://android-developers.googleblog.com/2017/03/update-your-app-to-take-advantage-of.html

Note: I've been unable to test so this fix is speculative.