implemented requested changes
1. fixed indentation. I added "set expandtab" to my .vimrc

2. use %f format instead of %e format for the numbers in the polar strincgs.
This ensures connected devices which don't use libc to convert
string to double don't choke

3. added "POV,C,NA" sentences to the protocol by which XCSoar tells the
connected device that a requested value has not been set by the user.

Example:
the connected device requests Real Polar and MacCread setting:
$POV,?,RPO,MC
XCSoar respose will be:
$POV,C,RPO,0.002867,-0.195457,4.272843*54
$POV,C,NA,MC*27

If the user has set the MacCread setting:
XCSoar respose will be:
$POV,C,RPO,0.002867,-0.195457,4.272843*54
$POV,C,MC,1.50*32

A detailed documentation of the NMEA protocol will be in the OpenVario Documentation.