new 12 InfoBox layout: 4 boxes X 3 rows
This layout of 12 InfoBoxes is meant to support tall and narrow displays.
It makes the boxes wider to allow bigger fonts. Taller screens can afford
more space for InfoBoxes without compromising the map area.
The intention is to improve readability of InfoBox title, comment and value.
in particular on tall and narrow displays.