scalable InfoBox title font
Introduce a user parameter "InfoBox title size" 50% ... 150% to scale
the font for the title and comment in the InfoBoxes.
The intention is to improve readability of InfoBox title and comment
in particular on tall and narrow displays. It allows the user to trade off
readability vs. length of text being displayed.