Repository: Add RepoFilemanager and persist repo file hahes
The RepoFileManager object provides an interface to interact with
a FileRepository obect. It also maintains a map of hashes to
determine if a file is up to date or not. These hashes are persisted
in a local file `local_hashes`.

This logic is moved out of the FileManager Dialog so it can also
be used elsewhere.