Add ability to set orientation of initial screens
The problem: currently the display orientation (aka rotation) is set in
the profile. XCSoar will rotate the screen according to
"DisplayOrientation" setting soon as profile is loaded. However, there are
several UIs that are shown before the profile is read. Notably, the
initial "Fly/Simulator" screen and the actual profile selector UI. XCSoar
would always render these screens in default "landscape" orientation.
This causes problems for some devices with fixed installation (such as
Openvario), on which these initial screens might appear "sideways".

The solution: add ability to detect the display orientation from the
OS environment and apply it before the first window is displayed.