Look/InfoBox: use FontScale() for the title font size
On large displays (tablets and computers), the InfoBox titles are
getting too large, because their size is scaled with the InfoBox
width, ignoring the physical size.

This commits introduces an upper bound for the title font size,
calculated using Layout::FontScale().  This function calculates the
size from a physical size, and applies the configured UI scaling
factor.  This way, users should get the same (physical) font size on
all displays, while obeying their preferences.

This is a better solution instead of
 https://github.com/XCSoar/XCSoar/pull/420