Implement FANET '#FNF' service (type 4) sentences decoding
This commit implements FANET driver for '#FNF' sentences of type 4(service) that contain current weather info and draws related wind arrow and info box with wind speed and wind gust to the screen.

You can test it with following sentence:
#FNF FD,FFFF,1,0,4,F,7a7f74406e450a12aeb7cb17c91601

Location is near Rijeka, Croatia -> latitude: 45.320333, longitude: 14.444006
Wind is coming from 245deg
Wind speed is 55km/h
Wind gust is 75km/h

Temperature, pressure, humidity and state of charge are also decoded but currently not presented to user in any way.

Fanet #FNF sentences and protocol specification can be found at:
https://github.com/3s1d/fanet-stm32/raw/master/fanet_module.pdf
https://github.com/3s1d/fanet-stm32/blob/master/Src/fanet/radio/protocol.txt

I had to to do a slight modification to the NMEAInfo::ExpireWallClock() so it doesnt invalidate(alive.Clear()) NMEAInfo object, after 10 seconds of inactivity on the device, if it contains alive weather stations so it can be merged by mergeThread and passed to MapWindow.