net/http/Request: disable TLS verification on Android
Since commit 3ec575628fc4a7, libcurl is initialised properly for the
first time on Android, and now all TLS connections fail because the CA
bundle could not be loaded.  Until we figure out how to use Android's
CA certificates with libcurl, this commit is a kludge to restore the
previous insecure but working behavior.