ActionInterface: send exchange command directly to device
The previous implementation swapped the frequencies by setting a new active frequency and a new standby frequency and forwarding them to a device if needed. 

XCSoar has no other use for radio frequencies than for interaction with external devices. A device should offer the possibility to accept a swap command directly, rather than indirectly being enforced. None of the current radio devices where an XCSoar driver exists appears to offer a public documentation that states that a direct swap command is accepted by it (except the Air Control Display). Furthermore, until this commit, the ExchangeRadioFrequencies action was unused in the code.

Due to the reasons stated above, the method is changed to send the ExchangeRadioFrequencies command to the devices directly.