build/thirdparty: build a full binutils/GCC toolchain for Kobo
The Kobo build was always problematic; glibc was bad because the glibc
version on the Kobo stock firmware was too old, and shipping our own
was too bloated.  Static linking with glibc breaks NSS modules.  Then
we migrated to Musl, which solved those problems, but caused new
problems because we now had to rebuild libstdc++ for the Musl ABI.
That libstdc++ version is sometimes not compatible with the GCC
version the user just happened to be using.

Instead of inventing more bad kludges, this commit attempts to do the
one right thing: not only build a C library from scratch, but also the
compiler (GCC including libstdc++) and binutils.  This way, developers
don't need to install an ARM cross compiler.