Protection: read the MergeThead::FirstRun() results during startup
This fixes a rare freeze bug in the airspace/waypoint dialog: in fly
mode, when no device is connected and the MergeThread never runs again
(after FirstRun()), AttitudeState::heading is uninitialised and may
contain a random value.  That value may lead to an endless loop in
Angle::AsBearing() via FormatBearing(), effectively freezing XCSoar.

The simple solution is to copy the DeviceBlackboard to the
InterfaceBlackboard again after MergeThread::FirstRun() has invoked
the BasicComputer for the first time.