android/DownloadUtil: download to getExternalFilesDir()
New Android versions restrict access to arbitrary storage directories;
the requestLegacyExternalStorage option (enabled in commit 1c2b1efa9b)
allows us to keep using XCSoarData for some time, but will eventually
be disabled.

However, on some Android versions, the DownloadManager refuses to
download to XCSoarData, even if requestLegacyExternalStorage is
enabled.  This means we need a solution now, and download to a
temporary file inside getExternalFilesDir().  After that, move that
temporary file to the final path.

Closes https://github.com/XCSoar/XCSoar/issues/549