android/AndroidManifest.xml: request the ACCESS_BACKGROUND_LOCATION permission
This is required since Android 11, or else XCSoar doesn't get any GPS
updates while running in background.

Documentation (confusing and contradictory):
 https://developer.android.com/about/versions/oreo/background-location-limits
 https://developer.android.com/training/location/permissions#request-location-access-runtime

Some discussion and clarifications:
 https://teddit.net/r/androiddev/comments/khe6z0/access_background_location_permission_declaration/

Apparently, it is not necessary to call
requestPermissions(ACCESS_BACKGROUND_LOCATION), and doing so would
automatically deny it; for this feature to work, the user has to check
the Android permission settings and change "allow while using the
application" to "always allow".

This is yet another sad Android usability regression.

Closes https://github.com/XCSoar/XCSoar/issues/618