Android/MainWindow: don't cancel simulator prompt on pause
Don't terminate the XCSoar process if the app gets paused; this allows
Android to keep the XCSoar process around.

As a side effect, this fixes the XCSoar shutdown if XCSoar is launched
into split screen mode, which occurs because Android pauses XCSoar
right after launching it, for reasons unknown to me.

Closes https://github.com/XCSoar/XCSoar/issues/674