InfoBoxes/Window: repaint custom content only if it has really changed
This commit adds a kind of serial number or digest number which
identifies the current contents; if that number doesn't change, the
buffer doesn't need to be redrawn.

This is mostly important on the Kobo when a fast vario is connected;
it will trigger lots of redraws, and the Kobo is very slow at updating
the screen.

Closes https://github.com/XCSoar/XCSoar/issues/297
Closes https://github.com/XCSoar/XCSoar/issues/665