Use task defaults when height ref is missing in task file
In task files, generated by external tools some task attributes might be
missing in the task xml file. When loading such tasks we want to keep
task default values, set in the XCSoar system settings.

For example, tasks, gnerated by soarscore.com are missing start and
finish height settings. This is on purpose, because this information is
not available on soaringspot.com, which is used as a source for the task
generation. Instead of putting some arbitrary values in the generated
task xml file, attributes, such as finish_min_height or
finish_min_height_ref are simply omitted. When XCSoar loads such a task
file, it correctly sets FinishConstraints.min_height to the value from
the task defaults, however min_height_ref gets always reset to "AGL".
When competition rules specify min. finish height as MSL, user has to
manually adjust this task property every time the task is loaded.

This fixes this issue by making GetHeightRef() behave more like
ConstDataNode.GetAttribute(): when attribute is not present in the node,
value is not modified.