Android/*Helper: check for Java exceptions again after ctor lookup
Turns out that JNIEnv::FindClass() only looks up the class, but does
not initialzie it; it does not run static initializers, for example.
Only the first method lookup will do that.  So to be sure that a Java
class is usable, we need to check for Java exceptions twice, or the
JVM will crash inside art::Thread::AssertNoPendingException().