build: Cross-build for Cubieboard supports DRI/KMS instead of Mali blob
Allow setting "ENABLE_MESA_KMS" for cross-compilation for Cubieboards
(usually Cubieboard 2 with A20, these are used for OpenVario).
DRI/KMS based EGL and OPENGL ES 2.0 is now supported by the open-source
LIMA kernel driver starting in 5.2, and user-space driver in Mesa.
Mesa with LIMA support is shipped out-of-the-box with distributions for
Armhf.

To cross-compile with KMS for Cubie enter:
make TARGET=CUBIE CUBIE=<path to the Cubieboard root directory> \
  ENABLE_MESA_KMS=y

OPENGL and GLES2 are automatically set.

For current Ubuntu versions (>= Hirsute)
"TARGET_LDFLAGS += -L/usr/arm-linux-gnueabihf/lib"
is required because the core C shared libraries must be pulled from the
cross-compiler package. The core libraries are not reachable on the
target root file system because they are sym-linked with an absoulte
path, not the current directory.

Signed-off-by: Kai Horstmann <horstmannkai@hotmail.com>