Topography/TopographyFileRenderer: move PaintPoint() to a separate loop
Decouple line/polygon drawing from point drawing, because the two use
very different (OpenGL) renderers.  Points are drawn by drawing a
simple texture, and lines/polygons use a vertex shader.  Switching
between the two modes was buggy, and fixing it is not worthwile; by
moving the two to a separate loop, better optimisations are possible
(but not implemented here).

Closes https://github.com/XCSoar/XCSoar/issues/702