support of WCH CH9102 USB to UART bridge controller
CH9102F is a pin-to-pin compatible replacement for CP2104
https://www.cnx-software.com/2021/09/14/ch9102f-a-replacement-for-cp2104-usb-to-uart-bridge/

It becomes very popular in year of 2021 since the price tag for CP2104 is growing up rapidly.

This PR contains VID and PID of CH9102 for the java USB filter to allow processing of the device.

The CH9102 does typically operate with generic USB CDC driver.