debug.mk: Allow minimal optimizations for Debug compile
A lot of warnings are activated including fancy ones like
-Wmissing-noreturn.
These warnings require data flow analysis which is not done by gcc with
-O0.

-Og does not re-arrange code, and still allows good debugging. It still
performs basic optimizations, and it performance wise much closer to
production builds.
More importantly it activates data flow analysis which enables the
compiler warnings based on it.

Signed-off-by: Kai Horstmann <horstmannkai@hotmail.com>