build: Forward C++ warning options to the LTO linker
Invoking g++ as linker with warning options has no effect for linking
machine code objects.
For LTO linking warning options become relevant because part of the
compilation is executed at link time.
In addition -Werror used for debug builds to treat any warning as error
is invoked for the linker to.

Signed-off-by: Kai Horstmann <horstmannkai@hotmail.com>