Fix startup for Clara
This adds a 0.1s delay in rcS file for Kobo to allow touch screen
and display to initialize properly.  this was specifically tested
for Clara HD (N249) but other models have had similar problems and
those models may now work better.

Closes #611
Closes #317
not sure if #515 is fixed because I don't have that model of Kobo