LocalPath: support Context.getExternalFilesDir()
Default to storing all data files in an app-specific directory.  This
is bad, because that directory will be deleted if you uninstall
XCSoar, and it cannot be shared between XCSoar-stable and
XCSoar-testing, but it's the only directory where Android apps have
full control starting with targetSdkVersion="30" (i.e. without option
"requestLegacyExternalStorage").