android: update targetSdkVersion to 30
This is required by Google Play since november 2021; this is the
reason we could not upload 7.21 there.

The reason we did not do this step so far is because this disables the
"requestLegacyExternalStorage" option (see commit 1c2b1efa9bf1190);
but users with existing XCSoar installations can preserve the option
with the new "preserveLegacyExternalStorage" option.