Android/Context: add missing "null" checks in getExternalFilesDirs() elements
According to
https://developer.android.com/reference/android/content/Context#getExternalFilesDirs(java.lang.String)
elements of the array returned by Context.getExternalFilesDirs() may
be null:

 "Some individual paths may be null if that shared storage is not
 currently available"

This has occurred on a Galaxy S4 with Android 5.0, a phone with an SD
card slot which I left empty.  This bug went unnoticed for a long
time, apparently because modern phones don't have SD card slots.