Do not show pseudo/virtual terminals
The list of TTYs is too long.
This improves the filtering
to only include devices that
are not pseudo/virtual terminals.
The resulting list should only contain
serial ports that are connected to sensor
hardware. Devices that have a symlink
generated by a udev rule will show up with
both the symlink and the original device name.