Add keylock function in lua, fixes for waypointlist
New function keylock to lock a keyboard for lua commands. This is
helpful for bigger keyboards like numpad in narrow cockpits.
It prevents undesired keyboard inputs.
Updated lua.rst

Fix for Waypoint:
Waypoint list dialog hung when it was opened for the second time.
Reason: There was a call to the uninitialized Listbox.