Enable debugging for Android Builds
Set android:debuggable dynamically in android.mk
If DEBUG=y, the flag will be set to true.
It the flag is set, the app will wait for the debugger. This is not
helpful for non debugging sessions.

Change the Loader.java and android.mk to work with -ns.so libraries
(containing symbols). This makes it much easier to debug in Android
Studio.

Fixed a warning from newer android sdk in MyServices.java

A more detailed description how to work with Android Studio and XCSoar
will follow later.